+++ api/src/main/java/muni/pa165/api/facade/UserFacade.java
@@ -1,0 +1,4 @@
+package muni.pa165.api.facade;
+
+public interface UserFacade {
+}
