+++ persistence/src/main/java/muni/pa165/persistence/enums/EventType.java
@@ -1,0 +1,11 @@
+package muni.pa165.persistence.enums;
+
+
+/**
+ * Event Type
+ *
+ * @author Usman Zia
+ */
+public enum EventType {
+    TENNIS_USER,LESSON,TOURNAMENT
+}
