+++ persistence/src/main/java/muni/pa165/persistence/enums/UserType.java
@@ -1,0 +1,12 @@
+package muni.pa165.persistence.enums;
+
+/**
+ * User role Type
+ *
+ * @author Muhammad Abdullah
+ */
+
+public enum UserType {
+    MANAGER,
+    TENNIS_USER
+}
