+++ pom.xml
@@ -11,6 +11,12 @@
             </plugin>
         </plugins>
     </build>
+    <modules>
+        <module>api</module>
+        <module>services</module>
+        <module>persistance</module>
+        <module>persistence</module>
+    </modules>
 
 
     <parent>
@@ -23,7 +29,7 @@
     <groupId>org.example</groupId>
     <artifactId>PA165-2021-Project</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
+    <packaging>pom</packaging>
 
 
     <properties>
@@ -169,5 +175,4 @@
                 <version>1</version>
             </dependency>
         </dependencies>
-
 </project>
