+++ services/src/main/java/muni/pa165/services/UserServiceImpl.java
@@ -1,0 +1,7 @@
+package muni.pa165.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserServiceImpl {
+}
