+++ services/src/main/java/muni/pa165/services/config/ServiceConfig.java
@@ -1,0 +1,4 @@
+package muni.pa165.services.config;
+
+public class ServiceConfig {
+}
