+++ services/src/main/java/muni/pa165/services/facade/UserFacadeImpl.java
@@ -1,0 +1,4 @@
+package muni.pa165.services.facade;
+
+public class UserFacadeImpl {
+}
