--- src/main/java/muni/pa165/PersistenceApplicationContext.java
@@ -1,67 +1,0 @@
-package muni.pa165;
-
-import org.hibernate.jpa.HibernatePersistenceProvider;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.instrument.classloading.InstrumentationLoadTimeWeaver;
-import org.springframework.instrument.classloading.LoadTimeWeaver;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-
-import javax.sql.DataSource;
-import java.util.Objects;
-
-@Configuration
-@EnableTransactionManagement
-@EnableJpaRepositories
-@ComponentScan(basePackages = "muni.pa165")
-public class PersistenceApplicationContext {
-
-    /**
-     * Enables automatic translation of exceptions to DataAccessExceptions.
-     */
-    @Bean
-    public PersistenceExceptionTranslationPostProcessor postProcessor() {
-        return new PersistenceExceptionTranslationPostProcessor();
-    }
-
-    @Bean
-    public JpaTransactionManager transactionManager() {
-        return new JpaTransactionManager(Objects.requireNonNull(entityManagerFactory().getObject()));
-    }
-
-    /**
-     * Starts up a container that emulates behavior prescribed in JPA spec for container-managed EntityManager
-     */
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
-        LocalContainerEntityManagerFactoryBean jpaFactoryBean = new LocalContainerEntityManagerFactoryBean();
-        jpaFactoryBean.setDataSource(db());
-        jpaFactoryBean.setLoadTimeWeaver(instrumentationLoadTimeWeaver());
-        jpaFactoryBean.setPersistenceProviderClass(HibernatePersistenceProvider.class);
-        return jpaFactoryBean;
-    }
-
-    @Bean
-    public LocalValidatorFactoryBean localValidatorFactoryBean() {
-        return new LocalValidatorFactoryBean();
-    }
-
-    @Bean
-    public LoadTimeWeaver instrumentationLoadTimeWeaver() {
-        return new InstrumentationLoadTimeWeaver();
-    }
-
-    @Bean
-    public DataSource db() {
-        EmbeddedDatabaseBuilder builder = new EmbeddedDatabaseBuilder();
-        return builder.setType(EmbeddedDatabaseType.DERBY).build();
-    }
-}
