--- src/main/java/muni/pa165/enums/EventType.java
@@ -1,11 +1,0 @@
-package muni.pa165.enums;
-
-
-/**
- * Event Type
- *
- * @author Usman Zia
- */
-public enum EventType {
-    TENNIS_USER,LESSON,TOURNAMENT
-}
