--- src/main/java/muni/pa165/enums/UserType.java
@@ -1,12 +1,0 @@
-package muni.pa165.enums;
-
-/**
- * User role Type
- *
- * @author Muhammad Abdullah
- */
-
-public enum UserType {
-    MANAGER,
-    TENNIS_USER
-}
